+++ assets/src/scripts/components/hydrograph/axes.js
@@ -1,7 +1,7 @@
 import { axisBottom, axisLeft } from 'd3-axis';
 import { createSelector } from 'reselect';
 import { DateTime } from 'luxon';
-import { wrap } from '../../utils';
+import { wrap, deltaDays } from '../../utils';
 import { getYTickDetails } from './domain';
 import { layoutSelector } from './layout';
 import { xScaleSelector, yScaleSelector } from './scales';
@@ -14,8 +14,15 @@ import { mediaQuery } from '../../utils';
 const FORMAT = {
     P7D: 'MMM dd',
     P30D: 'MMM dd',
-    P1Y: 'MMM yyyy'
+    P1Y: 'MMM yyyy',
+    custom: null
 };
+
+const setP7D = () => {
+    let date = tzStartDate.startOf('day');
+    let timePeriod = 'days';
+    let interval = 1;
+}
 
 /**
  * Generate the values for ticks to place on a hydrograph.
@@ -32,24 +39,53 @@ export const generateDateTicks = function(startDate, endDate, period, ianaTimeZo
     let date;
     let timePeriod;
     let interval;
+    let dateDiff;
+
+    const setP7D = () => {
+        date = tzStartDate.startOf('day');
+        timePeriod = 'days';
+        interval = 1;
+    };
+    const setP30D = () => {
+        date = tzStartDate.minus({days: tzStartDate.weekday}).startOf('day');
+        timePeriod = 'weeks';
+        interval = 1;
+    };
+    const setP1Y = () => {
+        date = tzStartDate.startOf('month');
+        timePeriod = 'months';
+        if (mediaQuery(config.USWDS_LARGE_SCREEN)) {
+            interval = 1;
+        } else {
+            interval = 2;
+        }
+    };
     switch (period) {
         case 'P7D':
-            date = tzStartDate.startOf('day');
-            timePeriod = 'days';
-            interval = 1;
+            setP7D();
             break;
         case 'P30D':
-            date = tzStartDate.minus({days: tzStartDate.weekday}).startOf('day');
-            timePeriod = 'weeks';
-            interval = 1;
+            setP30D();
             break;
         case 'P1Y':
-            date = tzStartDate.startOf('month');
-            timePeriod = 'months';
-            if (mediaQuery(config.USWDS_LARGE_SCREEN)) {
-                interval = 1;
+            setP1Y();
+            break;
+        case 'custom':
+            dateDiff = deltaDays(new Date(startDate), new Date(endDate));
+            if (dateDiff <= 7) {
+                setP7D();
+                FORMAT.custom = 'MMM dd';
+            } else if (7 < dateDiff && dateDiff <= 30) {
+                setP30D();
+                FORMAT.custom = 'MMM dd';
+            } else if (30 < dateDiff && dateDiff <= 365) {
+                setP1Y();
+                FORMAT.custom = 'MMM yyyy';
             } else {
-                interval = 2;
+                date = tzStartDate.startOf('month');
+                timePeriod = 'months';
+                interval = Math.ceil(dateDiff/365.25);
+                FORMAT.custom = 'MMM yyyy';
             }
             break;
         default: