+++ assets/src/scripts/selectors/time-series-selector.spec.js
@@ -319,6 +319,18 @@ describe('timeSeriesSelector', () => {
                                 }
                             }
                         }
+                    },
+                    'current:custom:00095': {
+                        notes: {
+                            requestDT: 1568729153803,
+                            'filter:timeRange': {
+                                mode: 'RANGE',
+                                interval: {
+                                    start: 1454738400000,
+                                    end: 1557896400000
+                                }
+                            }
+                        }
                     }
                 },
                 variables: TEST_VARS
@@ -353,6 +365,13 @@ describe('timeSeriesSelector', () => {
             expect(getRequestTimeRange('current', 'P30D', '00060')(TEST_DATA)).toEqual({
                 start: 1488348000000,
                 end: 1490763600000
+            });
+        });
+
+        it('should use the interval for a request with a custom RANGE', () => {
+            expect(getRequestTimeRange('current', 'custom', '00095')(TEST_DATA)).toEqual({
+                start: 1454738400000,
+                end: 1557896400000
             });
         });
     });